+++ dapps/reports/src/views/sections/TokenDetails.tsx
@@ -1,4 +1,4 @@
-import TokenDetailContainer from '@karpatkey-monorepo/reports/src/views/sections/TokenDetails/TokenDetailContainer'
+import TokenDetailContainer from '@karpatkey-monorepo/reports/src/views/sections/TokenDetailItems/TokenDetailContainer'
 import PaperSection from '@karpatkey-monorepo/shared/components/PaperSection'
 import BoxWrapperColumn from '@karpatkey-monorepo/shared/components/Wrappers/BoxWrapperColumn'
 import * as React from 'react'
